uint256 -> uint refactoring, work on Crowdsale

+++ contracts/ERC20.sol
@@ -2,13 +2,13 @@ pragma solidity ^0.4.11;
 
 // ERC: Token standard #20, https://github.com/ethereum/eips/issues/20
 contract ERC20 {
-	function totalSupply() constant returns (uint256 totalSupply);
-	function balanceOf(address _owner) constant returns (uint256 balance);
-	function transfer(address _to, uint256 _value) returns (bool success);
-	function transferFrom(address _from, address _to, uint256 _value) returns (bool success);
-	function approve(address _spender, uint256 _value) returns (bool success);
-	function allowance(address _owner, address _spender) constant returns (uint256 remaining);
+	function totalSupply() constant returns (uint totalSupply);
+	function balanceOf(address _owner) constant returns (uint balance);
+	function transfer(address _to, uint _value) returns (bool success);
+	function transferFrom(address _from, address _to, uint _value) returns (bool success);
+	function approve(address _spender, uint _value) returns (bool success);
+	function allowance(address _owner, address _spender) constant returns (uint remaining);
 
-	event Transfer(address indexed _from, address indexed _to, uint256 _value);
-	event Approval(address indexed _owner, address indexed _spender, uint256 _value);
+	event Transfer(address indexed _from, address indexed _to, uint _value);
+	event Approval(address indexed _owner, address indexed _spender, uint _value);
 }
